commit the Float->double patch for algebra so we have the weekend to sort out any build issues. Also this splits out the vector distribution generators and adds one for the surface of a box
git-svn-id: https://svn.salilab.org/imp/trunk@1547 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/algebra/include/Matrix2D.h
@@ -286,8 +286,8 @@ public:
   }
 
   //! Determinant (only for 2x2)
-  Float det() {
-    return (Float)((*this)(0, 0)*(*this)(1, 1) - (*this)(1, 0)*(*this)(0, 1));
+  double det() {
+    return (double)((*this)(0, 0)*(*this)(1, 1) - (*this)(1, 0)*(*this)(0, 1));
   }
 
 protected:
